migrating to separate slew_scan module
